+++ backend/lib/services/terminals/index.js
@@ -10,7 +10,7 @@ const _ = require('lodash')
 const hash = require('object-hash')
 const yaml = require('js-yaml')
 const config = require('../../config')
-const { encodeBase64 } = require('../../utils')
+const { getClusterCaData } = require('../shoots')
 const { cleanKubeconfig } = require('@gardener-dashboard/kube-config')
 const { Resources } = require('@gardener-dashboard/kube-client')
 
@@ -336,21 +336,6 @@ async function getTargetCluster ({ user, namespace, name, target, preferredHost,
     }
   }
   return targetCluster
-}
-
-async function getClusterCaData (client, { namespace, name }) {
-  try {
-    const configmap = await client.core.configmaps.get(namespace, `${name}.ca-cluster`)
-    return encodeBase64(configmap.data?.['ca.crt'])
-  } catch (err) {
-    // TODO(petersutter): Remove this fallback of reading the `<shoot-name>.ca-cluster` Secret when Gardener no longer reconciles it, presumably with Gardener v1.97.
-    if (isHttpError(err, 404)) {
-      const caCluster = await client.core.secrets.get(namespace, `${name}.ca-cluster`)
-      return caCluster.data['ca.crt']
-    } else {
-      throw err
-    }
-  }
 }
 
 async function getGardenTerminalHostCluster (client, { body }) {
